Extract shared message utilities and sanitize attachment file names
